#171 DatePattern now in a separate package
