+++ demo-v14/src/main/java/org/vaadin/miki/MainView.java
@@ -30,8 +30,8 @@ import org.vaadin.miki.markers.CanSelectText;
 import org.vaadin.miki.markers.HasDatePattern;
 import org.vaadin.miki.markers.HasLocale;
 import org.vaadin.miki.markers.WithNullValueOptionallyAllowed;
-import org.vaadin.miki.superfields.dates.DatePattern;
-import org.vaadin.miki.superfields.dates.DatePatterns;
+import org.vaadin.miki.shared.dates.DatePattern;
+import org.vaadin.miki.shared.dates.DatePatterns;
 import org.vaadin.miki.superfields.dates.SuperDatePicker;
 import org.vaadin.miki.superfields.dates.SuperDateTimePicker;
 import org.vaadin.miki.superfields.itemgrid.ItemGrid;
