+++ superfields/src/main/java/org/vaadin/miki/markers/WithDatePatternMixin.java
@@ -1,6 +1,6 @@
 package org.vaadin.miki.markers;
 
-import org.vaadin.miki.superfields.dates.DatePattern;
+import org.vaadin.miki.shared.dates.DatePattern;
 
 /**
  * Mixin interface to allow chaining {@link #setDatePattern(DatePattern)}.
