+++ superfields/src/main/java/org/vaadin/miki/superfields/dates/DatePatternDelegate.java
@@ -3,6 +3,7 @@ package org.vaadin.miki.superfields.dates;
 import com.vaadin.flow.component.AttachEvent;
 import com.vaadin.flow.component.Component;
 import org.vaadin.miki.markers.HasDatePattern;
+import org.vaadin.miki.shared.dates.DatePattern;
 
 import java.io.Serializable;
 import java.time.LocalDate;
